move LocationAutocomplete in components folder
